updated split func

+++ kaggle/dogs-vs-cats/split_data.py
@@ -1,25 +1,35 @@
 import os, shutil
 
-dataset_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data')
-train_dir = os.path.join(dataset_dir, 'train')
-small_train_dir = os.path.join(dataset_dir, 'small_train')
-val_dir = os.path.join(dataset_dir, 'validation')
-test_dir = os.path.join(dataset_dir, 'test')
 
-train_cats_dir = os.path.join(small_train_dir, 'cats')
-train_dogs_dir = os.path.join(small_train_dir, 'dogs')
+def mkne(directory):
+    '''
+    mkdir if not exists
+    '''
+    if not os.path.exists(directory):
+        os.makedirs(directory)
+    return directory
 
-val_cats_dir = os.path.join(val_dir, 'cats')
-val_dogs_dir = os.path.join(val_dir, 'dogs')
+dataset_dir = mkne(os.path.join(os.path.dirname(os.path.abspath(__file__)), 'data'))
+train_dir = mkne(os.path.join(dataset_dir, 'train'))
+small_train_dir = mkne(os.path.join(dataset_dir, 'small_train'))
+val_dir = mkne(os.path.join(dataset_dir, 'validation'))
+test_dir = mkne(os.path.join(dataset_dir, 'test'))
 
-test_cats_dir = os.path.join(test_dir, 'cats')
-test_dogs_dir = os.path.join(test_dir, 'dogs')
+train_cats_dir = mkne(os.path.join(small_train_dir, 'cats'))
+train_dogs_dir = mkne(os.path.join(small_train_dir, 'dogs'))
+
+val_cats_dir = mkne(os.path.join(val_dir, 'cats'))
+val_dogs_dir = mkne(os.path.join(val_dir, 'dogs'))
+
+test_cats_dir = mkne(os.path.join(test_dir, 'cats'))
+test_dogs_dir = mkne(os.path.join(test_dir, 'dogs'))
 
 def chdir(fnames, src_dir, dst_dir):
     for fn in fnames:
         src = os.path.join(src_dir, fn)
         dst = os.path.join(dst_dir, fn)
-        shutil.move(src, dst)
+        if os.path.exists(src) and not os.path.exists(dst):
+            shutil.move(src, dst)
 
 if __name__ == '__main__':
       chdir([f'cat.{i}.jpg' for i in range(1, 1000)], train_dir, train_cats_dir)
